introduce
/JDK 1.8 AND TOMCAT 9
This is a sign-in system for student.You can add coursetable so that student can choose course.Then i add searching and submit function.
Addmistrater can delete student's record and select students who doesn't sign-in.
ps. this is a jsp project.I am still learing about it ,so if you have a good idea or suggestion about it,welcome to share.